Refactor dataset standards; project helpers
The expectation for saved datasets has now been
changed to require that a column in a pickled dataframe
have byte vectors already created. This removes
the need for the user to specify a column for the
vector field.

The goal is to eventually have helper functions
in the library to make this easy.

Lastly, a bunch of project helpers and python project
standard items were added.