Change Search modifier and add getting started ex
The search property was nice for adding new modifiers
but left the vector param name opaque to the user. this
change will now make the user choose the param name
or work on the default.

Also added a getting started example that helps walk users
through the library interface and pandas reader