Llmcache ttl configuration (#28)
Enables TTL for loading data to Redis using RedisVL. This TTL is an
optional kwarg for the underlying `load` function. This is particularly
useful for clients and abstractions like SemanticCache that lean on
Redis support for ephemeral data.

Also adds a `clear` method to the LLM cache class that invalidates all
data from the cache, but does not disturb the index itself.