Add optional data load preprocessor hook (#67)
Another finding from working the arxiv example.. often you need to
unpack or edit a record before writing to the source. If you do this
before invoking redisvl, you add an additional loop, one that is
unnecessary in the end. So this allows devs to optionally add a
preprocessor method on load to call against each record.