Upgrade to use Pydantic 2.x.x (#101)
With the shift in the ecosystem to Pydantic 2.x.x, we need to get
`redisvl` to the same level.

Referenced here: https://github.com/RedisVentures/redisvl/issues/86

Fortunately, there is not a ton of complex usage of Pydantic; which
makes this easy! Changes include:

- Shifting the storage classes to use Pydantic to avoid manual (and
ugly) field validation. Now, the subclasses inherit from the base and
params are all typed as expected.
- Shifting the vectorizer classes to use Pydantic for the same reasons.
- Bump and pin pydantic version in the packge requirements to a
reasonable range `pydantic>=2.0.0,<3`
- use `v1` shim for now for safety