Add ability drop all keys from the index (#183)
From our chat today I'm pulling out this method from the semantic cache
and moving it into index. Rather can modify the `clear()` method to take
an optional list of keys it's a new method.

The goal is to stop the accidental calling of clear() with no args,
thinking it will clear no keys, but actually delete everything.