Use pydantic for cache entries and hits (#195)
This PR handles 2 things:

1) Uses pydantic for both `CacheEntry` and `CacheHit` models. It ended
up being better and cleaner to have two separate models.... one for what
we write, and the other for what we load.

2) Filtering. @justin-cechmanek I want your feedback on this. I landed
on a technique to allow for the user to define a list of
`filterable_fields` as part of the semantic cache class init. What this
does is give us a way to support arbitrary filters (scope, permissions,
tags, numerics.... anything) within reason. Then you can create any
`FilterExpression` and pass through at query time. This extends your
initial implementation.