Consider vectorizer the owner of dtype (#267)
The `dtype` argument to `SemanticCache`, `SemanticRouter`, and
`SemanticSessionManager` is intended to specify the vectorizer's data
type. However, the constructors for these classes allow passing in both
a vectorizer instance (with dtype already set) and `dtype`, in which
case, the vectorizer may have a different `dtype` than the `dtype` used
in the index schema.

This PR works on eliminating this possibility by adding validation and
treating the vectorizer as the true "owner" of `dtype`. We also begin
the deprecation process for standalone `dtype` arguments, guiding users
to pass in vectorizer instances instead if they require customizing the
vectorizer's dtype. Finally, we make vectorizer validation consistent
across these classes.