Upgrade to Pydantic V2 models (#281)
This PR introduces `pydantic ^2.0.0` models, validators, and techniques.
The implementation here improves performance of model construction and
validation. It also introduces best practices when possible related to
validators and serialization. Because RedisVL was already on pydantic
v2, this change should not impact any end users as we already required
this version and had been relying on the `v1` shim.