Validate EF_RUNTIME query parameter (#319)
The EF_RUNTIME parameter is only valid in queries that target vector
fields using the HNSW algorithm. If you run the same query against a
'flat' vector field, Redis will return an error. Because we have all the
information to prevent you from running the bad query, we can introduce
validation that raises a helpful error instead of relying on Redis error
handling.

---------

Co-authored-by: Tyler Hutcherson <tyler.hutcherson@redis.com>