Fix numpy version and remove scipy optional (#347)
Original goal of the task was to relax the forced `numpy` version
constraint that was causing issues in Python 3.11 environments where
numpy 2.x.x was required (like Google Colab)
![Screenshot 2025-06-03 at 1 58
50 PM](https://github.com/user-attachments/assets/d1f1bfca-6a81-4cad-8669-e379fca4ba75)

## Problem
However, I quickly ran into dependency chain issues with:
- `scipy` -- the scientific computing stack is a bit like a dinosaur
still
- `boto3` -- always an issue
- `ranx` -- not really sure why, but this one has 13+ separate
dependencies and includes a pin on `numba` that expects certain things
from `numpy` and then it gets easily circular

## Solution:
Working together with claude, I was able to resolve and rebuild the
environment and the lockfile with one major change.
- `ranx` is still an optional, but it's not an explicit extra.
- What does that mean? Literally only means that if devs are using
`ranx` they have to install it with `pip install ranx` instead of `pip
install redisvl[ranx]`

This could be some kind of limitation with the poetry version I was
using. But I spent 6 hours on this with Claude and probably spent 50$ in
credits on working through an onion of python environment. It now works
and our issues on google colab are gone. I also think the package ranges
are a bit more generous now.