disable progress bar by default (#363)
https://github.com/redis/redis-vl-python/issues/362

TLDR: TQDM progress bars are annoying when shown every time you embed
something.

The default should be to turn these off for embedding calls. 

Not all packages use tqdm and instead a logger. I don't think redisvl
should be in charge of setting loglevel and that should be developer
configured but we can make a better base decision.