Decode LangCache attribute values on retrieval (#437)
In #429, we found and fixed an issue with LangCache attribute values,
which is that some characters are not allowed (commas) and some are
accepted but silently fail during retrieval (slashes). Our solution was
to encode these characters when we persist them and at query time. This
PR decodes the values in retrieved data as well.