Avoid configuring root logger in RedisVL (#445)
Closes #443 

This PR updates the logger configuration helper for RedisVL to avoid
configuring the root logger via `logging.basicConfig`. That operation
was interfering with user-configured loggers, causing app logs to be
emitted multiple times and making it trickier to hide RedisVL logs when
running in production.

In its place, we now configure a `NullHandler` (mainly to avoid issues
with loggers not having any handlers configured). **This means that any
RedisVL usage that expects logs to have a specific format must now
configure that format for itself.**